backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.25
alembic==1.13.1
pydantic==2.6.3
python-dotenv==1.0.1
httpx==0.26.0
openai==1.12.0
pypdf==4.1.0
pandas==2.2.1
openpyxl==3.1.2
fpdf2==2.7.7
python-multipart==0.0.9
jinja2==3.1.3
loguru==0.7.2
beautifulsoup4==4.12.3
rapidfuzz==3.6.1
